test(dag): remove core dev dependency cycle
