+++ packages/dag-core/package.json
@@ -56,8 +56,6 @@
     "zod-to-json-schema": "^3.24.6"
   },
   "devDependencies": {
-    "@robota-sdk/dag-adapters-local": "workspace:*",
-    "@robota-sdk/dag-node": "workspace:*",
     "eslint": "^8.56.0",
     "rimraf": "^5.0.5",
     "tsup": "^8.0.1",
